requirements.txt
# Azure OpenAI and AI Services
openai>=1.0.0
azure-identity>=1.12.0

# PDF Processing
PyMuPDF>=1.23.0
pypdf>=3.17.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional: For advanced PDF text extraction
pdfplumber>=0.10.0
pytesseract>=0.3.10  # For OCR if needed

# Optional: For data export
openpyxl>=3.1.0
xlsxwriter>=3.1.0
